rename logic folder
