+++ src/web/ts/features/homepage/homepage.types.ts
